GPII-3882: Greatly simplified the way we execute files
